+++ src/data-table/DetailsBox.tsx
